requirements.txt
streamlit
qrcode
pillow
pyzbar
fpdf